WIP mark RTM's task as done, almost done.
The passing of parameters to the Python method following the VIM help
does not seem to be working as expected, so I found a workaround by
accessing the s: variable directly from the Python script.

Marking RTM's tasks as done has been successfully implemented. However,
there is spaghetti code that needs to be refactored and a few more
conditions that need to be improved in order to make the initial version
of integration with RTM ready for release.